requirements.txt
greenlet          3.0.1
pip               23.2.1
playwright        1.40.0
pyee              11.0.1
python-dotenv     1.0.0
setuptools        65.5.0
typing_extensions 4.8.0